Move `lib/furniture` and `lib/utilities` up a level
