+++ app/furniture/breakout_tables_by_jitsi/_in_room.html.erb
