+++ app/furniture/check_dropbox/_in_room.html.erb
