+++ app/furniture/check_dropbox/checks/index.html.erb
