+++ app/furniture/furniture_controller.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class FurnitureController < ::ApplicationController
+  prepend_view_path 'app/lib'
+end
