+++ app/furniture/markdown_text_block/_in_room.html.erb
