+++ app/furniture/videobridge_by_jitsi.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# Provides an iFramed Jitsi Meet to a {Room}.
+class VideobridgeByJitsi
+  include Placeable
+end
