+++ app/furniture/videobridge_by_jitsi/_in_room.html.erb
