--- app/lib/furniture/furniture_controller.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Furniture
-  class FurnitureController < ::ApplicationController
-    prepend_view_path 'app/lib'
-  end
-end
