--- app/lib/furniture/videobridge_by_jitsi.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Furniture
-  # Provides an iFramed Jitsi Meet to a {Room}.
-  class VideobridgeByJitsi
-    include Placeable
-  end
-end
