--- app/lib/utilities/null_utility.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Utilities
-  class NullUtility < Utility
-  end
-end
