+++ app/utilities/null_utility.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class NullUtility < Utility
+end
