+++ config/routes.rb
@@ -14,7 +14,7 @@ Rails.application.routes.draw do
     resources :rooms, only: %i[show edit update new create destroy] do
       resource :waiting_room, only: %i[show update]
       resources :furniture_placements, only: %i[create update destroy]
-      namespace :furniture do
+      resource :furniture, only: [] do
         Furniture.append_routes(self)
       end
     end
