+++ spec/furniture/check_dropbox/check_policy_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
 
-RSpec.describe Furniture::CheckDropbox::CheckPolicy, type: :policy do
+RSpec.describe CheckDropbox::CheckPolicy, type: :policy do
   let(:subject) { described_class }
   let(:space) { FactoryBot.build_stubbed(:space) }
-  let(:check) { instance_double(Furniture::CheckDropbox::Check, space: space) }
+  let(:check) { instance_double(CheckDropbox::Check, space: space) }
   context 'as a guest' do
     let(:actor) { Guest.new }
     permissions :show?, :update?, :edit?, :destroy?, :index? do
