+++ spec/furniture/check_dropbox/checks_request_spec.rb
@@ -10,9 +10,9 @@ RSpec.describe '/spaces/:space_id/rooms/:room_id/furniture/check_dropbox/checks'
   let(:neighbor) { FactoryBot.create(:person) }
   let(:space_member) { FactoryBot.create(:person, spaces: [space]) }
 
-  let(:fake_plaid) { instance_double(Utilities::Plaid) }
+  let(:fake_plaid) { instance_double(PlaidUtility) }
   before do
-    allow(Utilities::Plaid).to receive(:new).and_return(fake_plaid)
+    allow(PlaidUtility).to receive(:new).and_return(fake_plaid)
     allow(fake_plaid).to receive(:exchange_public_token) do |attrs|
       double(access_token: "Access Token from #{attrs[:public_token]}",
              item_id: "Item Id from #{attrs[:public_token]}")
