+++ spec/models/item_repository_spec.rb
@@ -4,9 +4,9 @@ require 'rails_helper'
 
 RSpec.describe ItemRepository do
   let(:item_repository) do
-    ItemRepository.new(type: Furniture::CheckDropbox::Check, item_records: item_records, space: space)
+    ItemRepository.new(type: CheckDropbox::Check, item_records: item_records, space: space)
   end
-  let(:type) { Furniture::CheckDropbox::Check }
+  let(:type) { CheckDropbox::Check }
   let(:item_records) { location.item_records }
   let(:location) { FactoryBot.create(:furniture_placement) }
   let(:space) { location.space }
@@ -15,7 +15,7 @@ RSpec.describe ItemRepository do
       item = item_repository.new(status: 'green')
 
       expect(item.status).to eql(:green)
-      expect(item).to be_a(Furniture::CheckDropbox::Check)
+      expect(item).to be_a(CheckDropbox::Check)
       expect(location.item_records).to include(item.item_record)
     end
   end
@@ -25,7 +25,7 @@ RSpec.describe ItemRepository do
       item = item_repository.create(status: 'green')
 
       expect(item.status).to eql(:green)
-      expect(item).to be_a(Furniture::CheckDropbox::Check)
+      expect(item).to be_a(CheckDropbox::Check)
       expect(location.item_records).to include(item.item_record)
       expect(item).to be_persisted
     end
