+++ spec/utilities/plaid_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Utilities::Plaid, type: :model do
+RSpec.describe PlaidUtility, type: :model do
   it { is_expected.to validate_presence_of :client_id }
   it { is_expected.to validate_presence_of :secret }
   it { is_expected.to validate_presence_of :version }
